+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 - Put Rust extension module binary under `build/lib.*` directory. [#150](https://github.com/PyO3/setuptools-rust/pull/150)
 - Fix `Exec` binding with console scripts. [#154](https://github.com/PyO3/setuptools-rust/pull/154)
 
+### Changed
+- Replace `semantic_version` with `packaging` to parse version specs. [#157](https://github.com/PyO3/setuptools-rust/pull/157)
+
 ## 0.12.1 (2021-03-11)
 ### Fixed
 - Fix some files unexpectedly missing from `sdist` command output. [#125](https://github.com/PyO3/setuptools-rust/pull/125)